fix odometer part, fix bugs

+++ freertos_code/utils.cpp
@@ -3,9 +3,8 @@
 #include "task.h"
 
 
-// TBD - move it to some shared place
 float secondsSinceLastTick(TickType_t lastTick) {
     TickType_t diffTicks = xTaskGetTickCount() - lastTick;
-    float diffSec = diffTicks / configTICK_RATE_HZ;
+    float diffSec = (float) diffTicks  / (float) configTICK_RATE_HZ ;
     return diffSec;
 };
